Fixed bug in RepositoryService.removeMessage().

+++ src/RepositoryService.js
@@ -237,10 +237,11 @@ const handleRequest = {
             const tag = bali.component('#' + tokens[0]);
             const version = bali.component(tokens[1]);
             const bag = await repository.readDocument(tag, version);
-            const message = await repository.removeMessage(tag, version);
-            const response = await engine.encodeResponse(parameters, bag, message, true);
-            if (response.statusCode !== 200) await repository.addMessage(tag, version, message);
-            return response;
+            var message;
+            if (bag) {
+                message = await repository.removeMessage(tag, version);
+            }
+            return await engine.encodeResponse(parameters, bag, message, true);
         }
     }
